Refactor PSU adapters: move SimAdapter to adapters.psu package
