+++ adapters/psu/__init__.py
@@ -1,0 +1,5 @@
+from .sim import SimAdapter
+
+__all__ = [
+    "SimAdapter",
+]
