+++ adapters/psu/visa.md
@@ -1,0 +1,3 @@
+# VISA Notes for PSU Adapters
+
+תיעוד ראשוני עבור שילוב VISA עתידי (PyVISA) עם ספקי כוח.
