+++ devices/psu/adapters/__init__.py
@@ -1,1 +1,2 @@
-from .sim_adapter import SimAdapter
+# Deprecated shim: prefer `from adapters.psu import SimAdapter`
+from adapters.psu import SimAdapter
