+++ devices/psu/adapters/visa.md
@@ -1,9 +1,1 @@
-# PSU VISA Adapter Commands
-
-| Command                | Description                       | Response Example |
-|------------------------|-----------------------------------|------------------|
-| MEAS:VOLT?             | Read output voltage               | "12.000"         |
-| MEAS:CURR?             | Read output current               | "0.500"          |
-| SOUR:VOLT <value>      | Set output voltage                | none             |
-| SOUR:CURR <value>      | Set current limit                  | none             |
-| OUTP ON|OFF            | Enable/disable output              | none             |
+This document moved to adapters/psu/visa.md
