+++ test/test_psu.py
@@ -9,7 +9,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")
 
 
 from core.exceptions import ConnectionError
-from devices.psu.adapters.sim_adapter import SimAdapter
+from adapters.psu import SimAdapter
 from devices.psu.yaml_config_loader import YamlPSUConfigLoader
 from devices.psu.strategy import VirtualPsuStrategy, RealPsuStrategy
 from devices.psu.PsuDevice import PSU
